rename controller back to server
