+++ meson.build
@@ -35,7 +35,7 @@ sources = [
   'src/key_bindings/key_binding_quit.cpp',
   'src/output.cpp',
   'src/view.cpp',
-  'src/controller.cpp',
+  'src/server.cpp',
   'src/main.cpp',
 ]
