+++ src/main.cpp
@@ -1,4 +1,4 @@
-#include "controller.h"
+#include "server.h"
 
 int main(int argc, char *argv[]) {
   Server server;
